Updates of player scores is now done correctly

+++ WeCharades/src/com/example/wecharades/model/DataController.java
@@ -6,6 +6,7 @@ import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
+import java.util.Currency;
 import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
@@ -73,6 +74,7 @@ public class DataController extends Observable implements Observer{
 			if(dbm.getMessage() == DBMessage.ERROR){
 				DatabaseException e = (DatabaseException) dbm.getData();
 				Log.d("DatabaseError", "Code: " + e.getCode() + "   Message: " + e.getMessage());
+				e.printStackTrace();
 				setChanged();
 				notifyObservers(new DCMessage(DCMessage.ERROR, e.prettyPrint()));
 			} else if(dbm.getMessage() == DBMessage.MESSAGE){
@@ -267,6 +269,7 @@ public class DataController extends Observable implements Observer{
 				m.putTurns(dbGame.getValue());
 				if(dbGame.getKey().isFinished() 
 						&& dbGame.getKey().getPlayer2().equals(getCurrentPlayer())){
+					updatePlayerStats(getCurrentPlayer(), dbGame.getKey());
 					db.removeGame(dbGame.getKey());
 					removeVideofromServer(dbGame.getKey());
 				}
@@ -278,7 +281,9 @@ public class DataController extends Observable implements Observer{
 				//Update local - Also check for finish an delete db from db.
 				m.putGame(dbGame.getKey());
 				m.putTurns(dbGame.getValue());
-				if(dbGame.getKey().isFinished()){
+				if(dbGame.getKey().isFinished() 
+						&& dbGame.getKey().getPlayer2().equals(getCurrentPlayer())){
+					updatePlayerStats(getCurrentPlayer(), dbGame.getKey());
 					db.removeGame(dbGame.getKey());
 					removeVideofromServer(dbGame.getKey());
 				}
@@ -392,28 +397,34 @@ public class DataController extends Observable implements Observer{
 		break;
 		}
 		if(game.isFinished()){ //Update player stats
-			TreeMap<Player, Integer> scoreMap = getGameScore(game);
-			Player p1 = turn.getRecPlayer(); //This assignment is random, but it doesn't matter
-			Player p2 = turn.getAnsPlayer(); 
-			int p1GS = scoreMap.get(p1);
-			int p2GS = scoreMap.get(p2);
-			int p1W = 0;
-			int p2W = 0;
-			int p1L = 0;
-			int p2L = 0;
-			int draw = 0;
-			if(p1GS > p2GS){
-				p1W++;
-			} else if(p2GS > p1GS){
-				p2W++;
-			} else{
-				draw++;
-			}
-			db.incrementPlayerStats(p1, p1GS, p1W, draw, p1L);
-			db.incrementPlayerStats(p2, p2GS, p2W, draw, p2L);
+			updatePlayerStats(getCurrentPlayer(), game);
 		}
 		db.updateGame(game);
 		m.putGame(game);
+	}
+	
+	/**
+	 * Updates a players scores in the database - this has do be done by the logged in player!
+	 * @param p - The player to update
+	 * @param g - The game to update score from
+	 */
+	private void updatePlayerStats(Player p, Game g){
+		TreeMap<Player, Integer> scoreMap = getGameScore(g);
+		Player p1 = p;
+		Player p2 = g.getOpponent(p);
+		int p1GS = scoreMap.get(p1);
+		int p2GS = scoreMap.get(p2);
+		int p1W = 0;
+		int p1L = 0;
+		int draw = 0;
+		if(p1GS > p2GS){
+			p1W++;
+		} else if(p2GS > p1GS){
+			p1L++;
+		} else{
+			draw++;
+		}
+		db.incrementPlayerStats(p1, p1GS, p1W, draw, p1L);
 	}
